requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
flask-wtf
flask-cors

flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.9.0

paho-mqtt
python-dotenv
passlib
requests

reportlab
openpyxl
xlsxwriter

pandas
numpy
matplotlib

pillow
qrcode
